refactor Domain classes, change imports, add application.properties files, add @PropertySource to Spring configuration
