+++ team_zero_waste_restaurant/src/main/java/lv/javaguru/java2/wasterestarant/config/RestaurantApplicationConfiguration.java
@@ -2,9 +2,10 @@ package lv.javaguru.java2.wasterestarant.config;
 
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.PropertySource;
 
 @Configuration
 @ComponentScan(basePackages = "lv.javaguru.java2.wasterestarant")
-
+@PropertySource(value = "classpath:application.properties")
 public class RestaurantApplicationConfiguration {
 }
