+++ team_zero_waste_restaurant/src/main/java/lv/javaguru/java2/wasterestarant/console_ui/GetRestaurantMenuUIAction.java
@@ -3,7 +3,7 @@ package lv.javaguru.java2.wasterestarant.console_ui;
 import lv.javaguru.java2.wasterestarant.core.requests.GetRestaurantMenuRequest;
 import lv.javaguru.java2.wasterestarant.core.responses.GetRestaurantMenuResponse;
 import lv.javaguru.java2.wasterestarant.core.services.GetRestaurantMenuService;
-import lv.javaguru.java2.wasterestarant.domain.Dish;
+import lv.javaguru.java2.wasterestarant.core.domain.Dish;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
