+++ team_zero_waste_restaurant/src/main/java/lv/javaguru/java2/wasterestarant/console_ui/order/GetAllOrdersUIAction.java
@@ -4,7 +4,7 @@ import lv.javaguru.java2.wasterestarant.console_ui.UIAction;
 import lv.javaguru.java2.wasterestarant.core.requests.order.GetAllOrdersRequest;
 import lv.javaguru.java2.wasterestarant.core.responses.order.GetAllOrdersResponse;
 import lv.javaguru.java2.wasterestarant.core.services.order.GetAllOrdersService;
-import lv.javaguru.java2.wasterestarant.domain.Order;
+import lv.javaguru.java2.wasterestarant.core.domain.Order;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
