+++ team_zero_waste_restaurant/src/main/java/lv/javaguru/java2/wasterestarant/console_ui/wishlist/GetWishListUIAction.java
@@ -4,7 +4,7 @@ import lv.javaguru.java2.wasterestarant.console_ui.UIAction;
 import lv.javaguru.java2.wasterestarant.core.requests.wishlist.GetWishlistRequest;
 import lv.javaguru.java2.wasterestarant.core.responses.wishlist.GetWishlistResponse;
 import lv.javaguru.java2.wasterestarant.core.services.wishlist.GetWishlistService;
-import lv.javaguru.java2.wasterestarant.domain.OrderItem;
+import lv.javaguru.java2.wasterestarant.core.domain.OrderItem;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
