+++ team_zero_waste_restaurant/src/main/java/lv/javaguru/java2/wasterestarant/core/domain/Dish.java
@@ -1,4 +1,4 @@
-package lv.javaguru.java2.wasterestarant.domain;
+package lv.javaguru.java2.wasterestarant.core.domain;
 
 import java.util.List;
 import java.util.Objects;
