+++ team_zero_waste_restaurant/src/main/java/lv/javaguru/java2/wasterestarant/core/domain/Product.java
@@ -1,4 +1,4 @@
-package lv.javaguru.java2.wasterestarant.domain;
+package lv.javaguru.java2.wasterestarant.core.domain;
 
 import java.util.Date;
 import java.util.Objects;
