+++ team_zero_waste_restaurant/src/main/java/lv/javaguru/java2/wasterestarant/core/responses/GetRestaurantMenuResponse.java
@@ -1,6 +1,6 @@
 package lv.javaguru.java2.wasterestarant.core.responses;
 
-import lv.javaguru.java2.wasterestarant.domain.Dish;
+import lv.javaguru.java2.wasterestarant.core.domain.Dish;
 
 import java.util.List;
