+++ team_zero_waste_restaurant/src/main/java/lv/javaguru/java2/wasterestarant/core/responses/dish/GetAllDishesResponse.java
@@ -1,7 +1,7 @@
 package lv.javaguru.java2.wasterestarant.core.responses.dish;
 
 import lv.javaguru.java2.wasterestarant.core.responses.CoreResponse;
-import lv.javaguru.java2.wasterestarant.domain.Dish;
+import lv.javaguru.java2.wasterestarant.core.domain.Dish;
 
 import java.util.List;
 //Elena
