+++ team_zero_waste_restaurant/src/main/java/lv/javaguru/java2/wasterestarant/core/responses/ingredient/SearchIngredientResponse.java
@@ -2,7 +2,7 @@ package lv.javaguru.java2.wasterestarant.core.responses.ingredient;
 
 import lv.javaguru.java2.wasterestarant.core.responses.CoreError;
 import lv.javaguru.java2.wasterestarant.core.responses.CoreResponse;
-import lv.javaguru.java2.wasterestarant.domain.Ingredient;
+import lv.javaguru.java2.wasterestarant.core.domain.Ingredient;
 
 import java.util.List;
