+++ team_zero_waste_restaurant/src/main/java/lv/javaguru/java2/wasterestarant/core/responses/product/AddProductResponse.java
@@ -1,7 +1,7 @@
 package lv.javaguru.java2.wasterestarant.core.responses.product;
 import lv.javaguru.java2.wasterestarant.core.responses.CoreError;
 import lv.javaguru.java2.wasterestarant.core.responses.CoreResponse;
-import lv.javaguru.java2.wasterestarant.domain.Product;
+import lv.javaguru.java2.wasterestarant.core.domain.Product;
 import java.util.List;
 
 public class AddProductResponse extends CoreResponse {
