+++ team_zero_waste_restaurant/src/main/java/lv/javaguru/java2/wasterestarant/core/responses/product/GetAllProductsResponse.java
@@ -1,6 +1,6 @@
 package lv.javaguru.java2.wasterestarant.core.responses.product;
 
-import lv.javaguru.java2.wasterestarant.domain.Product;
+import lv.javaguru.java2.wasterestarant.core.domain.Product;
 
 import java.util.List;
