+++ team_zero_waste_restaurant/src/main/java/lv/javaguru/java2/wasterestarant/core/responses/wishlist/GetWishlistResponse.java
@@ -1,7 +1,7 @@
 package lv.javaguru.java2.wasterestarant.core.responses.wishlist;
 
 import lv.javaguru.java2.wasterestarant.core.responses.CoreResponse;
-import lv.javaguru.java2.wasterestarant.domain.OrderItem;
+import lv.javaguru.java2.wasterestarant.core.domain.OrderItem;
 
 import java.util.List;
