+++ team_zero_waste_restaurant/src/main/java/lv/javaguru/java2/wasterestarant/core/services/dish/SearchDishesService.java
@@ -6,8 +6,9 @@ import lv.javaguru.java2.wasterestarant.core.requests.Paging;
 import lv.javaguru.java2.wasterestarant.core.requests.dish.SearchDishesRequest;
 import lv.javaguru.java2.wasterestarant.core.responses.CoreError;
 import lv.javaguru.java2.wasterestarant.core.responses.dish.SearchDishesResponse;
-import lv.javaguru.java2.wasterestarant.domain.Dish;
+import lv.javaguru.java2.wasterestarant.core.domain.Dish;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 
 import java.util.ArrayList;
@@ -18,6 +19,12 @@ import java.util.stream.Collectors;
 //Elena
 @Component
 public class SearchDishesService {
+
+    @Value("${search.ordering.enabled}")
+    private boolean orderingEnabled;
+    @Value("${search.paging.enabled}")
+    private boolean pagingEnabled;
+
     @Autowired
     private Database database;
     @Autowired
@@ -35,7 +42,7 @@ public class SearchDishesService {
     }
 
     private List<Dish> order(List<Dish> dishes, Ordering ordering) {
-        if (ordering != null) {
+        if (orderingEnabled && (ordering != null)) {
             if ((ordering.getOrderBy().equals("name")) && (ordering.getOrderDirection().equals("ASCENDING"))) {
                 return dishes.stream().sorted(Comparator.comparing(Dish::getName))
                         .collect(Collectors.toList());
@@ -91,7 +98,7 @@ public class SearchDishesService {
     }
 
     private List<Dish> paging(List<Dish> books, Paging paging) {
-        if (paging != null) {
+        if (pagingEnabled && (paging != null)) {
             int skip = (paging.getPageNumber() - 1) * paging.getPageSize();
             return books.stream()
                     .skip(skip)
