+++ team_zero_waste_restaurant/src/main/java/lv/javaguru/java2/wasterestarant/core/services/ingredient/SearchIngredientService.java
@@ -6,8 +6,9 @@ import lv.javaguru.java2.wasterestarant.core.requests.Ordering;
 import lv.javaguru.java2.wasterestarant.core.requests.Paging;
 import lv.javaguru.java2.wasterestarant.core.responses.CoreError;
 import lv.javaguru.java2.wasterestarant.core.responses.ingredient.SearchIngredientResponse;
-import lv.javaguru.java2.wasterestarant.domain.Ingredient;
+import lv.javaguru.java2.wasterestarant.core.domain.Ingredient;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 
 import java.util.ArrayList;
@@ -16,6 +17,11 @@ import java.util.List;
 import java.util.stream.Collectors;
 @Component
 public class SearchIngredientService {
+
+    @Value("${search.ordering.enabled}")
+    private boolean orderingEnabled;
+    @Value("${search.paging.enabled}")
+    private boolean pagingEnabled;
 
     @Autowired
     private Database database;
@@ -36,7 +42,7 @@ public class SearchIngredientService {
     }
 
     private List<Ingredient> order(List<Ingredient> ingredients, Ordering ordering) {
-        if (ordering != null) {
+        if (orderingEnabled && (ordering != null)) {
             Comparator<Ingredient> comparator = ordering.getOrderBy().equals("N")
                     ? Comparator.comparing(Ingredient::getIngredient)
                     : Comparator.comparing(Ingredient::getQuantity);
@@ -51,7 +57,7 @@ public class SearchIngredientService {
     }
 
     private List<Ingredient> paging(List<Ingredient> ingredients, Paging paging) {
-        if (paging != null) {
+        if (pagingEnabled && (paging != null)) {
             int skip = (paging.getPageNumber() - 1) * paging.getPageSize();
             return ingredients.stream()
                     .skip(skip)
