+++ team_zero_waste_restaurant/src/main/java/lv/javaguru/java2/wasterestarant/core/services/order/CreateNewOrderService.java
@@ -2,7 +2,7 @@ package lv.javaguru.java2.wasterestarant.core.services.order;
 
 import lv.javaguru.java2.wasterestarant.core.database.Database;
 import lv.javaguru.java2.wasterestarant.core.requests.order.CreateNewOrderRequest;
-import lv.javaguru.java2.wasterestarant.domain.Order;
+import lv.javaguru.java2.wasterestarant.core.domain.Order;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
