+++ team_zero_waste_restaurant/src/main/java/lv/javaguru/java2/wasterestarant/core/services/products/GetAllProductsService.java
@@ -3,7 +3,7 @@ package lv.javaguru.java2.wasterestarant.core.services.products;
 import lv.javaguru.java2.wasterestarant.core.database.Database;
 import lv.javaguru.java2.wasterestarant.core.requests.product.GetAllProductsRequest;
 import lv.javaguru.java2.wasterestarant.core.responses.product.GetAllProductsResponse;
-import lv.javaguru.java2.wasterestarant.domain.Product;
+import lv.javaguru.java2.wasterestarant.core.domain.Product;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
