+++ team_zero_waste_restaurant/src/main/java/lv/javaguru/java2/wasterestarant/core/services/products/SearchProductService.java
@@ -6,8 +6,9 @@ import lv.javaguru.java2.wasterestarant.core.requests.Paging;
 import lv.javaguru.java2.wasterestarant.core.requests.product.SearchProductRequest;
 import lv.javaguru.java2.wasterestarant.core.responses.CoreError;
 import lv.javaguru.java2.wasterestarant.core.responses.product.SearchProductResponse;
-import lv.javaguru.java2.wasterestarant.domain.Product;
+import lv.javaguru.java2.wasterestarant.core.domain.Product;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 
 import java.util.ArrayList;
@@ -17,6 +18,11 @@ import java.util.stream.Collectors;
 
 @Component
 public class SearchProductService {
+
+    @Value("${search.ordering.enabled}")
+    private boolean orderingEnabled;
+    @Value("${search.paging.enabled}")
+    private boolean pagingEnabled;
 
     @Autowired
     private Database database;
@@ -35,7 +41,7 @@ public class SearchProductService {
     }
 
     private List<Product> order(List<Product> products, Ordering ordering) {
-        if (ordering != null) {
+        if (orderingEnabled && (ordering != null)) {
             Comparator<Product> comparator = ordering.getOrderBy().equals("BBD")
                     ? Comparator.comparing(Product::getExpiryDate)
                     : Comparator.comparing(Product::getQuantity);
@@ -57,7 +63,7 @@ public class SearchProductService {
     }
 
     private List<Product> paging(List<Product> products, Paging paging) {
-        if (paging != null) {
+        if (pagingEnabled && (paging != null)) {
             int skip = (paging.getPageNumber() - 1) * paging.getPageSize();
             return products.stream()
                     .skip(skip)
