+++ team_zero_waste_restaurant/src/main/java/lv/javaguru/java2/wasterestarant/core/services/wishlist/GetWishlistService.java
@@ -3,7 +3,7 @@ package lv.javaguru.java2.wasterestarant.core.services.wishlist;
 import lv.javaguru.java2.wasterestarant.core.database.Database;
 import lv.javaguru.java2.wasterestarant.core.requests.wishlist.GetWishlistRequest;
 import lv.javaguru.java2.wasterestarant.core.responses.wishlist.GetWishlistResponse;
-import lv.javaguru.java2.wasterestarant.domain.OrderItem;
+import lv.javaguru.java2.wasterestarant.core.domain.OrderItem;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
