+++ team_zero_waste_restaurant/src/test/java/lv/javaguru/java2/wasterestarant/core/services/GetRestaurantMenuServiceTest.java
@@ -3,8 +3,7 @@ package lv.javaguru.java2.wasterestarant.core.services;
 import lv.javaguru.java2.wasterestarant.core.database.Database;
 import lv.javaguru.java2.wasterestarant.core.requests.GetRestaurantMenuRequest;
 import lv.javaguru.java2.wasterestarant.core.responses.GetRestaurantMenuResponse;
-import lv.javaguru.java2.wasterestarant.core.responses.dish.GetAllDishesResponse;
-import lv.javaguru.java2.wasterestarant.domain.Dish;
+import lv.javaguru.java2.wasterestarant.core.domain.Dish;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.InjectMocks;
