+++ team_zero_waste_restaurant/src/test/java/lv/javaguru/java2/wasterestarant/core/services/order/GetAllOrdersServiceTest.java
@@ -3,8 +3,8 @@ package lv.javaguru.java2.wasterestarant.core.services.order;
 import lv.javaguru.java2.wasterestarant.core.database.Database;
 import lv.javaguru.java2.wasterestarant.core.requests.order.GetAllOrdersRequest;
 import lv.javaguru.java2.wasterestarant.core.responses.order.GetAllOrdersResponse;
-import lv.javaguru.java2.wasterestarant.domain.Order;
-import lv.javaguru.java2.wasterestarant.domain.OrderItem;
+import lv.javaguru.java2.wasterestarant.core.domain.Order;
+import lv.javaguru.java2.wasterestarant.core.domain.OrderItem;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.InjectMocks;
