+++ team_zero_waste_restaurant/src/test/resources/application.properties
@@ -1,0 +1,2 @@
+search.ordering.enabled=true
+search.paging.enabled=true
